tunnelKitOriginal: replace UInt64 in WireGuardDataCount
minor file consolidation
