--- Sources/TunnelKitWireGuardAppExtension/WgStats.swift
@@ -1,8 +1,0 @@
-//
-//  WgStats.swift
-//
-//
-//  Created by Yevgeny Yezub on 16/11/23.
-//
-
-import Foundation
